- Moved conditional renders into separate components - Re-enabled expected functionality for testing
