+++ tdrs-frontend/src/components/RequestAccessForm/RequestAccessForm.jsx
@@ -4,6 +4,8 @@ import { useDispatch } from 'react-redux'
 import FormGroup from '../FormGroup'
 import STTComboBox from '../STTComboBox'
 import { requestAccess } from '../../actions/requestAccess'
+import JurisdictionSelector from './JurisdictionSelector'
+import RegionSelector from './RegionSelector'
 
 function RequestAccessForm({ user, sttList }) {
   const errorRef = useRef(null)
@@ -22,59 +24,17 @@ function RequestAccessForm({ user, sttList }) {
   const [jurisdictionType, setJurisdictionTypeInputValue] = useState('state')
   const isAMSUser = user?.email?.includes('@acf.hhs.gov')
 
-  const [regional, setRegional] = useState(false)
-  const regionError = 'At least one Region'
-  const regionKey = 'regions'
-  const regionsNames = [
-    'Boston',
-    'New York',
-    'Philadelphia',
-    'Atlanta',
-    'Chicago',
-    'Dallas',
-    'Kansas City',
-    'Denver',
-    'San Francisco',
-    'Seattle',
-  ]
-
-  const excludeRegions = (state) => {
-    const { regions, ...newState } = state
-    return newState
-  }
-
-  const handleRegionChange = (event, regionPK) => {
-    const { name, checked } = event.target
-    const { [name]: removedError, ...rest } = errors
-    const newProfileInfo = { ...profileInfo }
-    if (!checked && newProfileInfo.regions.has(regionPK)) {
-      newProfileInfo.regions.delete(regionPK)
-    } else {
-      newProfileInfo.regions.add(regionPK)
-    }
-
-    const error = validation(name, newProfileInfo.regions)
-
-    setErrors({
-      ...rest,
-      ...(error && { [name]: touched[name] && error }),
-    })
-    setProfileInfo({ ...newProfileInfo })
-  }
-
   const validation = (fieldName, fieldValue) => {
     const field = {
       firstName: 'First Name',
       lastName: 'Last Name',
-      stt: isAMSUser && 'A state, tribe, or territory',
-      regions: regionError,
+      stt: !isAMSUser && 'A state, tribe, or territory',
     }[fieldName]
 
     if (
-      (Boolean(field) &&
-        typeof fieldValue === 'string' &&
-        fieldValue.trim() === '') ||
-      (fieldValue instanceof Set && fieldValue.size === 0)
+      Boolean(field) &&
+      typeof fieldValue === 'string' &&
+      fieldValue.trim() === ''
     ) {
       return `${field} is required`
     }
@@ -182,54 +142,10 @@ function RequestAccessForm({ user, sttList }) {
           handleChange={handleChange}
           handleBlur={handleBlur}
         />
-        {isAMSUser && (
-          <div className="usa-form-group">
-            <fieldset className="usa-fieldset">
-              <legend className="usa-label text-bold">Jurisdiction Type</legend>
-              <div className="usa-radio">
-                <input
-                  className="usa-radio__input"
-                  id="state"
-                  type="radio"
-                  name="jurisdictionType"
-                  value="state"
-                  defaultChecked
-                  onChange={() => setJurisdictionType('state')}
-                />
-                <label className="usa-radio__label" htmlFor="state">
-                  State
-                </label>
-              </div>
-              <div className="usa-radio">
-                <input
-                  className="usa-radio__input"
-                  id="tribe"
-                  type="radio"
-                  name="jurisdictionType"
-                  value="tribe"
-                  onChange={() => setJurisdictionType('tribe')}
-                />
-                <label className="usa-radio__label" htmlFor="tribe">
-                  Tribe
-                </label>
-              </div>
-              <div className="usa-radio">
-                <input
-                  className="usa-radio__input"
-                  id="territory"
-                  type="radio"
-                  name="jurisdictionType"
-                  value="territory"
-                  onChange={() => setJurisdictionType('territory')}
-                />
-                <label className="usa-radio__label" htmlFor="territory">
-                  Territory
-                </label>
-              </div>
-            </fieldset>
-          </div>
+        {!isAMSUser && (
+          <JurisdictionSelector setJurisdictionType={setJurisdictionType} />
         )}
-        {jurisdictionType && isAMSUser && (
+        {jurisdictionType && !isAMSUser && (
           <div
             className={`usa-form-group ${
               errors.stt ? 'usa-form-group--error' : ''
@@ -244,108 +160,16 @@ function RequestAccessForm({ user, sttList }) {
             />
           </div>
         )}
-        {!isAMSUser && (
-          <>
-            <div className="usa-form-group">
-              <fieldset className="usa-fieldset">
-                <legend className="usa-label text-bold">
-                  Do you work for an OFA Regional Office?*
-                </legend>
-                <div className="usa-radio">
-                  <input
-                    className="usa-radio__input"
-                    id="regional"
-                    type="radio"
-                    name="regionalType"
-                    value="regional"
-                    onChange={() => {
-                      if (displayingError) {
-                        setTouched({ ...touched, regions: true })
-                        setErrors({
-                          ...errors,
-                          regions: `${regionError} is required`,
-                        })
-                      }
-                      setProfileInfo({ ...profileInfo, regions: new Set() })
-                      setRegional(true)
-                    }}
-                  />
-                  <label className="usa-radio__label" htmlFor="regional">
-                    Yes
-                  </label>
-                </div>
-                <div className="usa-radio">
-                  <input
-                    className="usa-radio__input"
-                    id="central"
-                    type="radio"
-                    name="regionalType"
-                    value="central"
-                    defaultChecked
-                    onChange={() => {
-                      setErrors(excludeRegions(errors))
-                      setTouched(excludeRegions(touched))
-                      setProfileInfo(excludeRegions(profileInfo))
-                      setRegional(false)
-                    }}
-                  />
-                  <label className="usa-radio__label" htmlFor="central">
-                    No
-                  </label>
-                </div>
-              </fieldset>
-            </div>
-            {regional && (
-              <div
-                className={`usa-form-group ${regionKey in errors ? 'usa-form-group--error' : ''}`}
-              >
-                <fieldset className="usa-fieldset">
-                  <legend className="usa-label text-bold">Region(s)*</legend>
-                  <div>
-                    Need help?&nbsp;
-                    <a
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      href="https://www.acf.hhs.gov/oro/regional-offices"
-                    >
-                      Lookup region by location.
-                    </a>
-                  </div>
-                  {regionKey in errors && (
-                    <span
-                      className="usa-error-message"
-                      id={`regions-error-message`}
-                    >
-                      {errors.regions}
-                    </span>
-                  )}
-                  {regionsNames.map((region, index) => {
-                    return (
-                      <div className="usa-checkbox">
-                        <input
-                          className={`usa-checkbox__input ${regionKey in errors ? 'usa-input--error' : ''}`}
-                          id={region}
-                          type="checkbox"
-                          name={regionKey}
-                          value={region}
-                          aria-required="true"
-                          onChange={(event) =>
-                            handleRegionChange(event, index + 1)
-                          }
-                        />
-                        <label
-                          className={`usa-checkbox__label ${'regions' in errors ? 'usa-input--error' : ''}`}
-                          htmlFor={region}
-                        >
-                          Region {index + 1} ({region})
-                        </label>
-                      </div>
-                    )
-                  })}
-                </fieldset>
-              </div>
-            )}
-          </>
+        {isAMSUser && (
+          <RegionSelector
+            setErrors={setErrors}
+            errors={errors}
+            setTouched={setTouched}
+            touched={touched}
+            setProfileInfo={setProfileInfo}
+            profileInfo={profileInfo}
+            displayingError={displayingError}
+          />
         )}
         <Button type="submit" className="width-full request-access-button">
           Request Access
